pyproject.toml
[tool.poetry]
name = "assistent"
version = "0.1.0"
description = ""
authors = ["Vasiliynemix <vasyanemix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
loguru = "^0.7.2"
redis = "^5.0.3"
asyncpg = "^0.29.0"
black = "^24.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
